250_aks_service_connector/app-python/requirements.txt
azure-storage-blob==12.18.3
azure-identity==1.15.0